+++ sdk/ml/azure-ai-ml/CHANGELOG.md
@@ -11,7 +11,9 @@
   - Added support to select firewall sku to used for provisioning azure firewall when FQDN rules are added in
   AllowOnlyApprovedOutbound mode. FirewallSku options are `Standard` or `Basic`, defaults to `Standard`
   - Update TLS version from 1.0 to 1.2
-  - Added support for Distillation jobs. Can be created by importing `disillation` from `azure.ai.ml.model_customization`
+  - Added support for Distillation jobs. Can be created by importing `disillation` from `azure.ai.ml.distillation`
+  - Added Workspace property `ProvisionNetworkNow` to trigger the provisioning of the managed VNet with the default
+    Options when creating a Workspace with the managed VNet enabled, or else it does nothing
 ### Bugs Fixed
   - Fix send email notification issue in model monitoring